code tidy, expose prefetch to values and env var
