+++ FFCDemoPaymentService/Messaging/MesssageConfig.cs
@@ -7,6 +7,7 @@ namespace FFCDemoPaymentService.Messaging
         public string PaymentQueueName { get; set; }
         public string MessageQueueHost { get; set; }
         public string MessageQueuePort { get; set; }
+        public string MessageQueuePreFetch { get; set; }
         public string MessageQueueUser { get; set; }
         public string MessageQueuePassword { get; set; }
     }
