+++ FFCDemoPaymentService/appsettings.json
@@ -15,6 +15,7 @@
     "PaymentQueueName": "payment",
     "MessageQueueHost": "myqueue.servicebus.windows.net",
     "MessageQueuePort": "5671",
+    "MessageQueuePreFetch": "1",
     "MessageQueueUser": "mquser",
     "MessageQueuePassword": "mqpasswd"
   }
